+++ examples/rust/opensk/src/env/user_presence.rs
@@ -12,13 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use alloc::rc::Rc;
-use core::cell::Cell;
-
 use opensk_lib::api::connection::RecvStatus;
 use opensk_lib::api::user_presence::{UserPresence, UserPresenceError, UserPresenceWaitResult};
 use opensk_lib::ctap::status_code::Ctap2StatusCode;
-use wasefire::{button, scheduling, timer, usb};
+use wasefire::{scheduling, timer, usb};
+
+use crate::BLINK_MS;
 
 pub(crate) fn init() -> Impl {
     Impl(None)
@@ -27,33 +26,31 @@ pub(crate) fn init() -> Impl {
 pub(crate) struct Impl(Option<State>);
 
 struct State {
-    presence: Rc<Cell<bool>>,
-    _button: button::Listener<ButtonHandler>,
+    touch: crate::touch::Touch,
     _blink: crate::blink::Blink,
 }
 
 impl UserPresence for Impl {
     fn check_init(&mut self) {
-        let presence = Rc::new(Cell::new(false));
-        let button = button::Listener::new(0, ButtonHandler::new(&presence)).unwrap();
-        let blink = crate::blink::Blink::new_ms(500);
-        self.0 = Some(State { presence, _button: button, _blink: blink });
+        let touch = crate::touch::Touch::new(None);
+        let blink = crate::blink::Blink::new_ms(BLINK_MS);
+        self.0 = Some(State { touch, _blink: blink });
     }
 
     fn wait_with_timeout(
         &mut self, packet: &mut [u8; 64], timeout_ms: usize,
     ) -> UserPresenceWaitResult {
-        let presence = match &self.0 {
-            Some(x) => &x.presence,
+        let touch = match &self.0 {
+            Some(x) => &x.touch,
             None => return Err(Ctap2StatusCode::CTAP2_ERR_VENDOR_INTERNAL_ERROR),
         };
         let timeout = timer::Timeout::new_ms(timeout_ms);
         let mut listener = usb::ctap::Listener::new(usb::ctap::Event::Read);
         while !usb::ctap::read(packet).unwrap() {
             scheduling::wait_until(|| {
-                presence.get() || timeout.is_over() || listener.is_notified()
+                touch.is_present() || timeout.is_over() || listener.is_notified()
             });
-            if presence.get() {
+            if touch.is_present() {
                 return Ok((Ok(()), RecvStatus::Timeout));
             }
             if timeout.is_over() {
@@ -66,24 +63,5 @@ impl UserPresence for Impl {
 
     fn check_complete(&mut self) {
         self.0 = None;
-    }
-}
-
-struct ButtonHandler {
-    presence: Rc<Cell<bool>>,
-}
-
-impl ButtonHandler {
-    fn new(presence: &Rc<Cell<bool>>) -> Self {
-        ButtonHandler { presence: presence.clone() }
-    }
-}
-
-impl button::Handler for ButtonHandler {
-    fn event(&self, state: button::State) {
-        match state {
-            button::State::Released => (),
-            button::State::Pressed => self.presence.set(true),
-        }
     }
 }